+++ src/shapes/mod.rs
@@ -1,6 +1,5 @@
-pub mod collection;
-pub mod collection_mutation;
 pub mod rendering;
+pub mod collection;
 
 use crate::drawable::{DrawType, Drawable};
 use graphics_shapes::circle::Circle;